ssx: pull out sharded_service_container from application
I want to pull some sharded services into cloud_topics::app rather than
having them all be owned by application. To that end, this pulls out all
of the utilities for constructing and shutting down sharded services so
they may be reused.
