+++ src/v/redpanda/BUILD
@@ -78,6 +78,7 @@ redpanda_cc_library(
         "//src/v/security/audit",
         "//src/v/ssx:abort_source",
         "//src/v/ssx:future_util",
+        "//src/v/ssx:sharded_service_container",
         "//src/v/ssx:thread_worker",
         "//src/v/ssx:watchdog",
         "//src/v/storage",
