+++ src/v/redpanda/application.h
@@ -65,6 +65,7 @@
 #include "rpc/fwd.h"
 #include "rpc/rpc_server.h"
 #include "security/fwd.h"
+#include "ssx/sharded_service_container.h"
 #include "ssx/watchdog.h"
 #include "storage/api.h"
 #include "storage/fwd.h"
@@ -75,7 +76,6 @@
 #include <seastar/core/app-template.hh>
 #include <seastar/core/metrics_registration.hh>
 #include <seastar/core/sharded.hh>
-#include <seastar/util/defer.hh>
 
 #include <memory>
 
@@ -91,7 +91,7 @@ inline const auto redpanda_start_time{
   std::chrono::duration_cast<std::chrono::milliseconds>(
     std::chrono::system_clock::now().time_since_epoch())};
 
-class application {
+class application : public ssx::sharded_service_container {
 public:
     int run(int, char**);
 
@@ -220,14 +220,6 @@ public:
     }
 
 private:
-    // First warning timeout
-    static constexpr auto short_shutdown_warning_timeout = 15s;
-    // Time after which we will print the warning about the service shutdown
-    // taking too long.
-    static constexpr auto long_shutdown_warning_timeout = 120s;
-    using deferred_actions
-      = std::deque<ss::deferred_action<std::function<void()>>>;
-
     // Constructs and starts the services required to provide cryptographic
     // algorithm support to Redpanda
     void wire_up_and_start_crypto_services();
@@ -273,160 +265,6 @@ private:
 
     ss::shared_ptr<kafka::datalake_usage_api> make_datalake_usage_aggregator();
 
-    // Stop the service.
-    // The method should be invoked in the ss::thread context.
-    template<class Service>
-    void stop_service(
-      Service& s,
-      ss::sstring name,
-      std::optional<ss::sstring> next_to_stop = std::nullopt) {
-        // This watchdog is triggered after short period of time (30s). It
-        // adds message to the log that service is taking a long time to
-        // shutdown on INFO level.
-        ssx::watchdog short_wd(short_shutdown_warning_timeout, [this, name] {
-            vlog(
-              _log.info,
-              "Service {} is taking more than {} seconds to shut down.",
-              name,
-              std::chrono::duration_cast<std::chrono::seconds>(
-                short_shutdown_warning_timeout)
-                .count());
-        });
-        // This watchdog is triggered after long period of time. This indicates
-        // a bug (most likely).
-        ssx::watchdog long_wd(long_shutdown_warning_timeout, [this, name] {
-            vlog(
-              _log.info,
-              "Service {} is taking more than {} seconds to shut down!",
-              name,
-              std::chrono::duration_cast<std::chrono::seconds>(
-                long_shutdown_warning_timeout)
-                .count());
-        });
-        if (next_to_stop.has_value()) {
-            vlog(
-              _log.info,
-              "Stopping {}, ..next to shutdown is {}",
-              name,
-              *next_to_stop);
-        } else {
-            vlog(_log.info, "Stopping {}", name);
-        }
-        s.stop().get();
-        if (!next_to_stop.has_value()) {
-            vlog(_log.info, "Stopped {}", name);
-        }
-    }
-
-    template<class T>
-    static ss::sstring service_type_name() {
-        if constexpr (detail::is_specialization_of_v<T, std::unique_ptr>) {
-            return service_type_name<typename T::element_type>();
-        } else if constexpr (detail::is_specialization_of_v<T, std::vector>) {
-            return fmt::format(
-              "std::vector<{}>", service_type_name<typename T::value_type>());
-        }
-        return ss::pretty_type_name(typeid(T));
-    }
-
-    // Shutdown the service with custom method.
-    // The method should be invoked in the ss::thread context.
-    template<class Service, class StopFunc>
-    void shutdown_with_watchdog(
-      Service& s,
-      StopFunc stop_func,
-      const ss::sstring& name = service_type_name<Service>(),
-      std::source_location location = std::source_location::current()) {
-        auto start_watchdog = [&name, this](
-                                std::chrono::milliseconds timeout,
-                                ss::log_level log_level) {
-            return ssx::watchdog(timeout, [this, timeout, &name, log_level] {
-                vlogl(
-                  _log,
-                  log_level,
-                  "Service {} is taking more than {} seconds to shutdown",
-                  name,
-                  timeout / 1s);
-            });
-        };
-        // This watchdog is triggered after short period of time (30s). It
-        // adds message to the log that service is taking a long time to
-        // shutdown on INFO level.
-        ssx::watchdog short_wd = start_watchdog(
-          short_shutdown_warning_timeout, ss::log_level::info);
-        // This watchdog is triggered after long period of time. This indicates
-        // a bug (most likely).
-        ssx::watchdog long_wd = start_watchdog(
-          long_shutdown_warning_timeout, ss::log_level::error);
-
-        vlog(
-          _log.info,
-          "Shutting down: {} at {}:{}",
-          name,
-          location.file_name(),
-          location.line());
-        stop_func(s).get();
-        vlog(
-          _log.info,
-          "Shutdown completed: {} started at {}:{}",
-          name,
-          location.file_name(),
-          location.line());
-    }
-
-    /**
-     * @brief Construct service boilerplate.
-     *
-     * Construct the given service s, calling start with the given arguments
-     * and set up a shutdown callback to stop it.
-     *
-     * Returns the future from start(), typically you'll call get() on it
-     * immediately to wait for creation to complete.
-     *
-     * @return the future returned by start()
-     */
-    template<typename Service, typename... Args>
-    ss::future<> construct_service(ss::sharded<Service>& s, Args&&... args) {
-        auto name = ss::pretty_type_name(typeid(Service));
-        _deferred.emplace_back(
-          [this, &s, name, next_to_stop = _last_constructed_service_name] {
-              stop_service(s, name, next_to_stop);
-          });
-        _last_constructed_service_name = ss::sstring(name);
-        return s.start(std::forward<Args>(args)...);
-    }
-
-    template<typename Service, typename... Args>
-    void construct_single_service(std::unique_ptr<Service>& s, Args&&... args) {
-        auto name = ss::pretty_type_name(typeid(Service));
-        s = std::make_unique<Service>(std::forward<Args>(args)...);
-        _deferred.emplace_back(
-          [this, &s, name, next_to_stop = _last_constructed_service_name] {
-              stop_service(*s, name, next_to_stop);
-              s.reset();
-          });
-        _last_constructed_service_name = name;
-    }
-
-    template<typename Service, typename... Args>
-    ss::future<>
-    construct_single_service_sharded(ss::sharded<Service>& s, Args&&... args) {
-        auto name = ss::pretty_type_name(typeid(Service));
-        auto f = s.start_single(std::forward<Args>(args)...);
-        _deferred.emplace_back(
-          [this, &s, name, next_to_stop = _last_constructed_service_name] {
-              stop_service(s, name, next_to_stop);
-          });
-        _last_constructed_service_name = name;
-        return f;
-    }
-
-    template<typename ShardedComponent>
-    auto ref_to_local(ss::sharded<ShardedComponent>& component) {
-        return ss::sharded_parameter(
-          [&component] { return std::ref(component.local()); });
-    }
-
     void setup_metrics();
     void setup_public_metrics();
     void setup_internal_metrics();
@@ -451,7 +289,6 @@ private:
     std::optional<kafka::client::configuration> _schema_reg_client_config;
     std::optional<kafka::client::configuration> _audit_log_client_config;
     ss::sharded<scheduling_groups_probe> _scheduling_groups_probe;
-    ss::logger _log;
 
     std::optional<config::binding<bool>> _abort_on_oom;
 
@@ -493,10 +330,6 @@ private:
 
     ss::sharded<kafka::data::rpc::local_service> _kafka_data_rpc_service;
     ss::sharded<kafka::data::rpc::client> _kafka_data_rpc_client;
-
-    // run these first on destruction
-    deferred_actions _deferred;
-    std::optional<ss::sstring> _last_constructed_service_name;
 
     ss::sharded<aggregate_metrics_watcher> _aggregate_metrics_watcher;
